refactor: AfternoteDetailUi.kt 분리 및 컴포넌트 패키지 구조 개선
- `AfternoteDetailUi.kt` 파일에 포함되어 있던 공통 UI 컴포넌트들을 `shared.detail.components` 패키지로 분리 및 개별 파일로 이동
    - `InfoCard.kt`, `ReceiversCard.kt`, `DeleteConfirmDialog.kt`, `EditDropdownMenu.kt` 생성
- 컴포넌트 분리에 따른 프로젝트 내 임포트 경로 업데이트
- `GalleryDetailScreen`, `MemorialGuidelineDetailScreen`, `SocialNetworkDetailScreen` 등 상세 화면의 참조 수정
- `ReceiverAfterNoteMainScreen`, `AfternoteEditorReceiverList` 등 관련 화면의 참조 수정
